move query mapper to column root
